use parameter.read()
