+++ examples/tests/invalid/struct_from_parameter.test.w
@@ -11,7 +11,7 @@ let app = nodeof(this).app;
 // technically we never enforced the parameters must have been
 // provided
 
-MyParams.fromJson(app.parameters.json());
+MyParams.fromJson(app.parameters.read());
 //Error: unable to parse MyParams:
 // - instance requires property "bar"
 // - instance requires property "baz"
